Fix Lock Info functionality
Sometime after 4.0.9 was released, SmartThings made a change to their
default Z-Wave Lock Device Handler.  In this update the Poll function
was changed and reportAllCodes was added to an if statement that does
not occur when the lock data refresh is triggered.

This is due to “cmds” only being set if the conditions of the lock
state or update time is met in the previous if/else statement.

Because of this the lock data does not get refreshed in real time.

Removing reportAllCodes(state) out of the if statement and separating
it out by itself allows real time updating like previous versions of
the smart app.