Updates unlocked, unlocking, and unknown states to use amber background color
Updates the background color to that of the SmartSense open/close door
sensor device handler from SmartThings, which is an amber color to
denote the state of when the sensor is open. The amber color provides a
better visual indicator of caution when locks are either unlocked or
the state is unknown.